+++ APITours/Controllers/PaisesController.cs
@@ -59,6 +59,11 @@ namespace APITours.Controllers
             {
                 return NotFound();
             }
+            var tourExistente = await _context.Tours.AnyAsync(t => t.idPais == id);
+            if (tourExistente)
+            {
+                return BadRequest(new { mensaje = "No se puede eliminar el país porque está asociado a un tour." });
+            }
             _context.Pais.Remove(pais);
             await _context.SaveChangesAsync();
             return NoContent();
